Move Named and Classified interfaces to presentation module
