--- gaphor/diagram/abc.py
@@ -1,20 +1,0 @@
-"""
-A collection of some (abstract) base types, or marker interfaces,
-that can be used to identify groups of similar items.
-"""
-
-
-class Named:
-    """
-    Marker for any named presentations.
-    """
-
-    pass
-
-
-class Classified(Named):
-    """
-    Marker for Classified presentations.
-    """
-
-    pass
