+++ gaphor/diagram/actions/action.py
@@ -4,8 +4,7 @@ Action diagram item.
 
 from gaphor import UML
 from gaphor.UML.modelfactory import stereotypes_str
-from gaphor.diagram.abc import Named
-from gaphor.diagram.presentation import ElementPresentation
+from gaphor.diagram.presentation import ElementPresentation, Named
 from gaphor.diagram.support import represents
 from gaphor.diagram.shapes import Box, EditableText, Text, draw_boundry
