+++ gaphor/diagram/actions/activitynodes.py
@@ -13,8 +13,7 @@ from gaphas.geometry import Rectangle, distance_line_point
 
 from gaphor import UML
 from gaphor.UML.modelfactory import stereotypes_str
-from gaphor.diagram.abc import Named
-from gaphor.diagram.presentation import ElementPresentation
+from gaphor.diagram.presentation import ElementPresentation, Named
 from gaphor.diagram.support import represents
 from gaphor.diagram.shapes import Box, IconBox, EditableText, Text, draw_boundry
