+++ gaphor/diagram/actions/partition.py
@@ -9,8 +9,7 @@ TODO: partition can be resized only horizontally or vertically, therefore
 
 from gaphor import UML
 from gaphor.UML.modelfactory import stereotypes_str
-from gaphor.diagram.abc import Named
-from gaphor.diagram.presentation import ElementPresentation
+from gaphor.diagram.presentation import ElementPresentation, Named
 from gaphor.diagram.support import represents
 from gaphor.diagram.text import VerticalAlign
 from gaphor.diagram.shapes import Box, EditableText, Text, draw_highlight
