+++ gaphor/diagram/classes/association.py
@@ -20,8 +20,7 @@ from gaphas.state import reversible_property
 
 from gaphor import UML
 from gaphor.UML.modelfactory import stereotypes_str
-from gaphor.diagram.abc import Named
-from gaphor.diagram.presentation import LinePresentation
+from gaphor.diagram.presentation import LinePresentation, Named
 from gaphor.diagram.shapes import (
     Box,
     EditableText,
