+++ gaphor/diagram/components/connector.py
@@ -102,8 +102,7 @@ interfaces are connectable elements.
 
 from gaphor import UML
 from gaphor.UML.modelfactory import stereotypes_str
-from gaphor.diagram.abc import Named
-from gaphor.diagram.presentation import LinePresentation
+from gaphor.diagram.presentation import LinePresentation, Named
 from gaphor.diagram.shapes import Box, EditableText, Text
 from gaphor.diagram.support import represents
 from operator import attrgetter
