+++ gaphor/diagram/interactions/message.py
@@ -52,8 +52,7 @@ from gaphas.util import path_ellipse
 
 from gaphor import UML
 from gaphor.UML.modelfactory import stereotypes_str
-from gaphor.diagram.presentation import LinePresentation
-from gaphor.diagram.abc import Named
+from gaphor.diagram.presentation import LinePresentation, Named
 from gaphor.diagram.shapes import Box, EditableText, Text
 from gaphor.diagram.text import middle_segment
