+++ gaphor/diagram/profiles/extension.py
@@ -8,8 +8,7 @@ ExtensionItem -- Graphical representation of an association.
 
 from gaphor import UML
 from gaphor.UML.modelfactory import stereotypes_str
-from gaphor.diagram.abc import Named
-from gaphor.diagram.presentation import LinePresentation
+from gaphor.diagram.presentation import LinePresentation, Named
 from gaphor.diagram.shapes import Box, EditableText, Text
 from gaphor.diagram.support import represents
