+++ gaphor/diagram/states/transition.py
@@ -4,8 +4,7 @@ State transition implementation.
 
 from gaphor import UML
 from gaphor.UML.modelfactory import stereotypes_str
-from gaphor.diagram.presentation import LinePresentation
-from gaphor.diagram.abc import Named
+from gaphor.diagram.presentation import LinePresentation, Named
 from gaphor.diagram.shapes import Box, EditableText, Text, draw_arrow_tail
 from gaphor.diagram.support import represents
