+++ gaphor/diagram/usecases/actor.py
@@ -5,9 +5,8 @@ Actor item classes.
 from math import pi
 
 from gaphor import UML
-from gaphor.diagram.abc import Classified
 from gaphor.UML.modelfactory import stereotypes_str
-from gaphor.diagram.presentation import ElementPresentation
+from gaphor.diagram.presentation import ElementPresentation, Classified
 from gaphor.diagram.shapes import Box, IconBox, EditableText, Text
 from gaphor.diagram.support import represents
